score_service/requirements.txt
Flask==2.3.2
requests
psycopg2-binary  # Driver para PostgreSQL
redis          # Cliente para Redis
google-generativeai # API de Gemini
# Para el score (ejemplo):
sentence-transformers